Implement first functional version
No resource handling yet, as I think there is a problem with the
NaturalTransformation, taking the Future as a value and not a call by
name parameter.

Also implemented error handling for the REST service through meow-mtl,
though not working with returning Streams from the PrimeService, so I
will do a little refactor.